Upgrade bitswap-tunning to testground-go-sdk 0.2.3
