+++ bitswap-tuning/go.mod
@@ -20,6 +20,6 @@ require (
 	github.com/multiformats/go-multiaddr v0.2.0
 	github.com/multiformats/go-multihash v0.0.10
 	github.com/pkg/errors v0.9.1
-	github.com/testground/sdk-go v0.1.1
+	github.com/testground/sdk-go v0.2.3
 	golang.org/x/sync v0.0.0-20200317015054-43a5402ce75a
 )
